backend/requirements.txt
flask
uagents
google-generativeai
reportlab
weasyprint
python-dotenv
requests
